Ajuste varios
- Eliminar un metodo de la clase Pedido que a la final nunca se utilizó
- Se añade el setText(null) en GuiRegistrarPostre
- Se mueve el metodo de agragar al jcomboBox del cliente de la GuiRegistrarCliente.
- Se cambia el atributo telefono a tipo long

+++ src/software/Cliente.java
@@ -4,12 +4,12 @@ public class Cliente {
 
 	private String nombreCliente;
 	private String direccion;
-	private int telefono;
+	private long telefono;
 	private String sexo;
 	private int edad;
 
 	// Constructor explicito
-	public Cliente(String nombreCliente, String direccion, int telefono, String sexo, int edad) {
+	public Cliente(String nombreCliente, String direccion, long telefono, String sexo, int edad) {
 		this.nombreCliente = nombreCliente;
 		this.direccion = direccion;
 		this.telefono = telefono;
